+++ projects/js-packages/publicize-components/src/social-store/resolvers.js
@@ -1,13 +1,8 @@
 import { store as editorStore } from '@wordpress/editor';
-import { setAutoConversionSettings } from './actions/auto-conversion-settings';
 import { setConnections } from './actions/connection-data';
 import { setJetpackSettings } from './actions/jetpack-settings';
 import { setSocialImageGeneratorSettings } from './actions/social-image-generator-settings';
-import {
-	fetchJetpackSettings,
-	fetchSocialImageGeneratorSettings,
-	fetchAutoConversionSettings,
-} from './controls';
+import { fetchJetpackSettings, fetchSocialImageGeneratorSettings } from './controls';
 
 /**
  * Yield actions to get the Jetpack settings.
@@ -46,24 +41,6 @@ export function* getSocialImageGeneratorSettings() {
 }
 
 /**
- * Yield actions to get the Auto Conversion settings.
- *
- * @yields {object} - an action object.
- * @returns {object} - an action object.
- */
-export function* getAutoConversionSettings() {
-	try {
-		const settings = yield fetchAutoConversionSettings();
-		if ( settings ) {
-			return setAutoConversionSettings( settings.jetpack_social_autoconvert_images );
-		}
-	} catch ( e ) {
-		// TODO: Add proper error handling here
-		console.log( e ); // eslint-disable-line no-console
-	}
-}
-
-/**
  * Resolves the connections from the post.
  *
  * @returns {Function} Resolver
@@ -84,6 +61,5 @@ export function getConnections() {
 export default {
 	getJetpackSettings,
 	getSocialImageGeneratorSettings,
-	getAutoConversionSettings,
 	getConnections,
 };