+++ projects/packages/my-jetpack/_inc/components/product-cards-section/protect-card/index.tsx
@@ -1,36 +1,9 @@
-import { __ } from '@wordpress/i18n';
 import { useCallback, type FC } from 'react';
-import useProduct from '../../../data/products/use-product';
-import useAnalytics from '../../../hooks/use-analytics';
 import ProductCard from '../../connected-product-card';
 import ProtectValueSection from './protect-value-section';
 
-const ProtectCard: FC< { admin: boolean } > = ( { admin } ) => {
-	const { recordEvent } = useAnalytics();
+const ProtectCard: FC< { admin: boolean; recommendation?: boolean } > = props => {
 	const slug = 'protect';
-	const { detail } = useProduct( slug );
-	const { isPluginActive, hasPaidPlanForProduct: hasProtectPaidPlan } = detail || {};
-
-	/**
-	 * Called when secondary "View" button is clicked.
-	 */
-	const onViewButtonClick = useCallback( () => {
-		recordEvent( 'jetpack_myjetpack_product_card_manage_click', {
-			product: slug,
-		} );
-	}, [ recordEvent ] );
-
-	const shouldShowSecondaryButton = useCallback(
-		() => isPluginActive && ! hasProtectPaidPlan,
-		[ hasProtectPaidPlan, isPluginActive ]
-	);
-
-	const viewButton = {
-		href: 'admin.php?page=jetpack-protect',
-		label: __( 'View', 'jetpack-my-jetpack' ),
-		onClick: onViewButtonClick,
-		shouldShowButton: shouldShowSecondaryButton,
-	};
 
 	// This is a workaround to remove the Description from the product card. However if we end
 	// up needing to remove the Description from additional cards in the future, we might consider
@@ -39,10 +12,9 @@ const ProtectCard: FC< { admin: boolean } > = ( { admin } ) => {
 
 	return (
 		<ProductCard
-			admin={ admin }
+			{ ...props }
 			slug={ slug }
 			upgradeInInterstitial={ true }
-			secondaryAction={ viewButton }
 			Description={ noDescription }
 		>
 			<ProtectValueSection />