+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/breve/store/selectors.ts
@@ -37,20 +37,24 @@ export function getDisabledFeatures( state: BreveState ) {
 
 // Suggestions
 
-export function getBlockMd5( state: BreveState, feature: string, blockId: string ) {
-	return state.suggestions?.[ feature ]?.[ blockId ]?.md5 ?? '';
+export function getBlockMd5( state: BreveState, blockId: string ) {
+	return state.suggestions?.[ blockId ]?.md5 ?? '';
 }
 
 export function getSuggestionsLoading(
 	state: BreveState,
 	{ feature, id, blockId }: { feature: string; id: string; blockId: string }
 ) {
-	return state.suggestions?.[ feature ]?.[ blockId ]?.[ id ]?.loading;
+	return state.suggestions?.[ blockId ]?.[ feature ]?.[ id ]?.loading;
 }
 
 export function getSuggestions(
 	state: BreveState,
 	{ feature, id, blockId }: { feature: string; id: string; blockId: string }
 ) {
-	return state.suggestions?.[ feature ]?.[ blockId ]?.[ id ]?.suggestions;
+	return state.suggestions?.[ blockId ]?.[ feature ]?.[ id ]?.suggestions;
+}
+
+export function getIgnoredSuggestions( state: BreveState, { blockId }: { blockId: string } ) {
+	return state.suggestions?.[ blockId ]?.ignored;
 }