+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.58.0] - 2024-08-01
+### Changed
+- Fixup versions [#38612]
+
+### Removed
+- Removed the unused code for image auto-conversion from social store [#38609]
+- Social | Removed the media auto-conversion UI [#38497]
+
+### Fixed
+- Improved broken connection messaging for non-admins [#38639]
+- Social | Fixed and improved media auto conversion notices [#38499]
+
+## [0.57.0] - 2024-07-29
+### Added
+- Added mentioning of Manual Sharing [#38411]
+
+### Fixed
+- Fixed broken connection notices to make them more helpful [#38450]
+
 ## [0.56.2] - 2024-07-22
 ### Fixed
 - Fixed double request issue and simplifed refresh logic [#38350]
@@ -788,6 +807,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[0.58.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.57.0...v0.58.0
+[0.57.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.56.2...v0.57.0
 [0.56.2]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.56.1...v0.56.2
 [0.56.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.56.0...v0.56.1
 [0.56.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.55.1...v0.56.0